Bugfix: make_edges_live crash on unavailable edges
In the main function (edge_manage, line 117), a set of all edges
(including unavailable edges) called EdgeState is created in the object
EdgeManage.

Within the EdgeManage object another EdgeState set is created (this time
inside a DecisionMaker object), but without the unavailable edges (lines
155-162 of edgemanage.py).

In the last loop of make_edges_live function (around line 372), the loop
was looping over *all* edges, and consequently was asking the
DecisionMaker object for the health status of unavailable edges. These
are not in its array of EdgeState objects, and so it was crashing with a
KeyError exception.

This fix skips unavailable edges in that last loop to avoid the problem.

Signed-off-by: KheOps <kheops@ceops.eu>