Add configurable test web server
This commit adds a test web server which will show different performance
characteristics based on a configuration of a set of edge servers.

This web server uses the last octet of the request IP address as the
edge ID. This edge ID is then used to identify the target edge configuration from the config file.