Add testing mode and port number configuration options
The edge_manage testing webserver runs on port 5000, but edge_mange
currently only connected to port 80 or 443. I have added a
`testobject['port']` configuration option.

The testing server uses the HTTP Host header to determine which edge is being
requested. I have patched `edgemanage` to send the requested IP in the
host header when the `testobject['testing']` option is set to True.