Add basic integration tests where all edges responding
This commit adds the first integration test which makes request against
the testing web server. Much of scaffolding is now complete and it will
be much easier to add integration tests for different edge cases in
future.

In this test all edges respond immediately and so all measurements should
return a healthy response.