Replace ProcessPoolExecutor with ThreadPoolExecutor
We have been using the `ProcessPoolExecutor` for running HTTP fetches
to the edges asynchronously. The `ProcessPoolExecutor` spins out each
request to a different subprocess which adds a lot of extra overhead. It's
meant to be used for running CPU intensive tasks on separate CPU cores.

Our HTTP requests are all IO-bound see we should be using a ThreadPoolExecutor
instead. This has a much lower overhead and should be faster when starting to make requests too.

The `ProcessPoolExecutor` was also using the number of CPU's available
as the number of workers. On a VM this was probably quite low and would
unnecessarily limit the speed of requests.

I have now set the number of async workers to `10` and made it an option in
the configuration file. This number is just a guess so we may need to tweak it
up or down in the future.