Always return at least some edges for the DNS servers
Even when all edges are unresponsive, it is better to return slow edges
than to not return any edges at all from the DNS servers.

When all edges fail edge_manage will now reselect the last live set of
edges. These edges will be slow and failing but it's the best we can do
if edge health gets this bad.