Make HTTP requests more consistent and remove TLS validation warning
Different HTTP requests were being sent for the initial request than
for any retried HTTP requests. Retry requests were always initiated
without TLS certificate verification which was resulting in a lot of
noise in the edgemanage log file.

This commit uses the the same TLS verification option for all requests.
It also disables the InsecureRequestWarning to remove the noise from
the log files.