Immediately mark edge as failed when it exceeds TIMEOUT
A bug was introduced in decisionmaker.py in ef2c941. The average response
time was being used before checking if the latest fetch exceeded the
TIMEOUT.

This would mark an offline but previously working edge as `pass_average`
and would never mark it as failed as the average would approach but never
reach the TIMEOUT value.

The fix is to check the TIMEOUT value early before checking the fetch
average.

Fixes bug introduced in ef2c941.